fixes an issue where inferno-server was bundling inferno core
