+++ benchmarks/uibench-nonkeyed/app.js
@@ -4,7 +4,7 @@
 	/* (flags, type, props, children, key, ref, noNormalise) */
 	var createVNode = Inferno.createVNode;
 
-	uibench.init('Inferno', '1.0.0-beta33 -non-keyed *dev*');
+	uibench.init('Inferno', '1.0.0-beta34 -non-keyed *dev*');
 
 	var treeLeafProps = { className: 'TreeLeaf' };
