+++ package.json
@@ -2,7 +2,7 @@
   "name": "inferno-build",
   "private": true,
   "license": "MIT",
-  "version": "1.0.0-beta33",
+  "version": "1.0.0-beta34",
   "author": {
     "name": "Dominic Gannaway",
     "email": "dg@domgan.com"
