+++ packages/inferno-compat/package.json
@@ -1,6 +1,6 @@
 {
   "name": "inferno-compat",
-  "version": "1.0.0-beta33",
+  "version": "1.0.0-beta34",
   "license": "MIT",
   "description": "Provides a compatibility with React codebases",
   "author": {
@@ -21,12 +21,12 @@
     "user interfaces"
   ],
   "dependencies": {
-    "inferno": "1.0.0-beta33",
+    "inferno": "1.0.0-beta34",
     "inferno-transition-group": "1.1.0",
     "rc-css-transition-group": "2.1.4"
   },
   "devDependencies": {
     "proptypes": "0.14.3",
-    "inferno-component": "1.0.0-beta33"
+    "inferno-component": "1.0.0-beta34"
   }
 }
