+++ packages/inferno-mobx/package.json
@@ -1,6 +1,6 @@
 {
 	"name": "inferno-mobx",
-	"version": "1.0.0-beta33",
+	"version": "1.0.0-beta34",
 	"license": "MIT",
 	"description": "Official Inferno bindings for Mobx",
 	"author": {
@@ -24,7 +24,7 @@
 	],
 	"repository": "https://github.com/trueadm/inferno",
 	"dependencies": {
-		"inferno": "1.0.0-beta33",
+		"inferno": "1.0.0-beta34",
 		"mobx": "*"
 	}
 }
