+++ packages/inferno-redux/README.md
@@ -7,7 +7,7 @@ Inferno Redux passes `context.store` value to each component.
 ## Install
 
 ```
-npm install inferno-redux@beta33
+npm install inferno-redux@beta34
 ```
 
 ## Contents
