+++ packages/inferno-redux/package.json
@@ -1,6 +1,6 @@
 {
   "name": "inferno-redux",
-  "version": "1.0.0-beta33",
+  "version": "1.0.0-beta34",
   "license": "MIT",
   "description": "Official Inferno bindings for Redux",
   "author": {
@@ -23,9 +23,9 @@
   ],
   "repository": "https://github.com/trueadm/inferno",
   "dependencies": {
-    "inferno": "1.0.0-beta33",
-    "inferno-component": "1.0.0-beta33",
-    "inferno-create-element": "1.0.0-beta33",
+    "inferno": "1.0.0-beta34",
+    "inferno-component": "1.0.0-beta34",
+    "inferno-create-element": "1.0.0-beta34",
     "redux": "*"
   }
 }
