+++ packages/inferno-test-utils/package.json
@@ -1,6 +1,6 @@
 {
   "name": "inferno-test-utils",
-  "version": "1.0.0-beta33",
+  "version": "1.0.0-beta34",
   "license": "MIT",
   "description": "Provides test utilities to help with creating and maintaining tests for Inferno applications",
   "author": {
@@ -20,7 +20,7 @@
     "user interfaces"
   ],
   "dependencies": {
-    "inferno": "1.0.0-beta33"
+    "inferno": "1.0.0-beta34"
   },
   "main": "inferno-test-utils.js",
   "repository": "https://github.com/trueadm/inferno"
