+++ packages/inferno/package.json
@@ -1,6 +1,6 @@
 {
   "name": "inferno",
-  "version": "1.0.0-beta33",
+  "version": "1.0.0-beta34",
   "license": "MIT",
   "description": "An extremely fast, isomorphic JavaScript library for building modern user interfaces",
   "author": {
